+++ internal/RendererSIMD.h
@@ -279,7 +279,7 @@ void Ray::NS::RendererSIMD<DimX, DimY>::RenderScene(const std::shared_ptr<SceneB
 
     if (cam.pass_flags & OutputSH) {
         temp_buf_.ResetSampleData(rect);
-        for (size_t i = 0; i < secondary_rays_count; i++) {
+        for (int i = 0; i < secondary_rays_count; i++) {
             const auto &r = p.secondary_rays[i];
 
             simd_ivec<S> x = r.xy >> 16,
